requirements.txt
Pillow==9.5.0
numpy==1.25.0
python-dotenv==1.0.0
pandas==2.0.3
joblib==1.3.1
fastapi==0.100.1
uvicorn==0.23.1
pydantic==2.1.1
scikit-learn==1.6.1
psycopg2-binary==2.9.10